feat(cli): enhance JSON output support for list/show commands and add integration tests
